backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pandas
python-dotenv
bcrypt
psycopg2-binary
scikit-learn
